+++ src/counter.hpp
@@ -170,7 +170,7 @@ public:
   FF outer_count();
   uint64_t get_tstamp() const { return tstamp; }
   uint64_t get_tstamp(int32_t lev) const {
-    if (dec_level() < lev) return std::numeric_limits<uint32_t>::max();
+    if (dec_level() < lev) return std::numeric_limits<uint64_t>::max();
     return decisions[lev].tstamp;
   }
 
@@ -234,7 +234,7 @@ private:
   // Computing LBD (lbd == 2 means "glue clause")
   vector<uint64_t> lbd_helper;
   uint64_t lbd_helper_flag = 0;
-  template<class T2> uint32_t calc_lbd(const T2& lits);
+  template<class T2> uint8_t calc_lbd(const T2& lits);
 
   // Clause adding
   void simple_preprocess();
@@ -729,9 +729,9 @@ inline vector<Lit>::iterator Counter::top_declevel_trail_begin() {
 }
 
 template<class T2>
-uint32_t Counter::calc_lbd(const T2& lits) {
+uint8_t Counter::calc_lbd(const T2& lits) {
   lbd_helper_flag++;
-  uint32_t nblevels = 0;
+  uint8_t nblevels = 0;
   for(const auto& l: lits) {
     if (val(l) == X_TRI) {nblevels++;continue;}
     int lev = var(l).decision_level;
