+++ src/statistics.hpp
@@ -153,14 +153,14 @@ public:
            + sum_extra_bytes;
   }
 
-  void incorporate_cache_store(const uint32_t& extra_bytes, const uint32_t comp_nvars) {
+  void incorporate_cache_store(const uint64_t& extra_bytes, const uint32_t comp_nvars) {
     sum_extra_bytes += extra_bytes;
     sum_cache_store_sizes += comp_nvars;
     num_cached_comps++;
     total_num_cached_comps++;
   }
 
-  void incorporate_cache_erase(const uint32_t extra_bytes){
+  void incorporate_cache_erase(const uint64_t extra_bytes){
     sum_extra_bytes -= extra_bytes;
     num_cached_comps--;
   }
@@ -185,7 +185,7 @@ public:
 
   double cache_miss_rate() const {
     if(num_cache_look_ups == 0) return 0.0;
-    return (num_cache_look_ups - num_cache_hits)
+    return (double)(num_cache_look_ups - num_cache_hits)
         / (double) num_cache_look_ups;
   }
